Allow *_alter* to edit extended flag banks.
Block random mob selection from picking mobs in areas that are closed, locked to the player, or marked as no_get random

Block fading into locked areas.

+++ script_commands.c
@@ -4665,24 +4665,24 @@ SCRIPT_CMD(scriptcmd_alterobj)
 			return;
 		}
 	} else {
-		int *ptr = NULL;
+		long *ptr = NULL;
 
-		if(!str_cmp(field,"cond"))				ptr = (int*)&obj->condition;
-		else if(!str_cmp(field,"cost"))			{ ptr = (int*)&obj->cost; min_sec = 5; }
-		else if(!str_cmp(field,"extra"))		{ ptr = (int*)obj->extra; bank = extra_flagbank; sec_flags[1] = sec_flags[2] = sec_flags[3] = 5; }
-		else if(!str_cmp(field,"fixes"))		{ ptr = (int*)&obj->times_allowed_fixed; min_sec = 5; }
-		else if(!str_cmp(field,"level"))		{ ptr = (int*)&obj->level; min_sec = 5; }
-		else if(!str_cmp(field,"repairs"))		ptr = (int*)&obj->times_fixed;
-		else if(!str_cmp(field,"tempstore1"))	ptr = (int*)&obj->tempstore[0];
-		else if(!str_cmp(field,"tempstore2"))	ptr = (int*)&obj->tempstore[1];
-		else if(!str_cmp(field,"tempstore3"))	ptr = (int*)&obj->tempstore[2];
-		else if(!str_cmp(field,"tempstore4"))	ptr = (int*)&obj->tempstore[3];
-		else if(!str_cmp(field,"tempstore5"))	ptr = (int*)&obj->tempstore[4];
-		else if(!str_cmp(field,"timer"))		ptr = (int*)&obj->timer;
-		else if(!str_cmp(field,"type"))			{ ptr = (int*)&obj->item_type; flags = type_flags; min_sec = 7; }
-		else if(!str_cmp(field,"wear"))			{ ptr = (int*)&obj->wear_flags; flags = wear_flags; }
-		else if(!str_cmp(field,"wearloc"))		{ ptr = (int*)&obj->wear_loc; flags = wear_loc_flags; }
-		else if(!str_cmp(field,"weight"))		ptr = (int*)&obj->weight;
+		if(!str_cmp(field,"cond"))				ptr = (long*)&obj->condition;
+		else if(!str_cmp(field,"cost"))			{ ptr = (long*)&obj->cost; min_sec = 5; }
+		else if(!str_cmp(field,"extra"))		{ ptr = (long*)obj->extra; bank = extra_flagbank; sec_flags[1] = sec_flags[2] = sec_flags[3] = 5; }
+		else if(!str_cmp(field,"fixes"))		{ ptr = (long*)&obj->times_allowed_fixed; min_sec = 5; }
+		else if(!str_cmp(field,"level"))		{ ptr = (long*)&obj->level; min_sec = 5; }
+		else if(!str_cmp(field,"repairs"))		ptr = (long*)&obj->times_fixed;
+		else if(!str_cmp(field,"tempstore1"))	ptr = (long*)&obj->tempstore[0];
+		else if(!str_cmp(field,"tempstore2"))	ptr = (long*)&obj->tempstore[1];
+		else if(!str_cmp(field,"tempstore3"))	ptr = (long*)&obj->tempstore[2];
+		else if(!str_cmp(field,"tempstore4"))	ptr = (long*)&obj->tempstore[3];
+		else if(!str_cmp(field,"tempstore5"))	ptr = (long*)&obj->tempstore[4];
+		else if(!str_cmp(field,"timer"))		ptr = (long*)&obj->timer;
+		else if(!str_cmp(field,"type"))			{ ptr = (long*)&obj->item_type; flags = type_flags; min_sec = 7; }
+		else if(!str_cmp(field,"wear"))			{ ptr = (long*)&obj->wear_flags; flags = wear_flags; }
+		else if(!str_cmp(field,"wearloc"))		{ ptr = (long*)&obj->wear_loc; flags = wear_loc_flags; }
+		else if(!str_cmp(field,"weight"))		ptr = (long*)&obj->weight;
 
 		if(!ptr) return;
 
@@ -4895,7 +4895,7 @@ SCRIPT_CMD(scriptcmd_alterroom)
 	ROOM_INDEX_DATA *room;
 	WILDS_DATA *wilds;
 
-	int *ptr = NULL;
+	long *ptr = NULL;
 	int16_t *sptr = NULL;
 	char **str;
 	bool allow_empty = false;
@@ -5043,23 +5043,23 @@ SCRIPT_CMD(scriptcmd_alterroom)
 		return;
 	}
 
-	if(!str_cmp(field,"flags"))				{ ptr = (int*)room->room_flag; bank = room_flagbank; }
-	else if(!str_cmp(field,"light"))		{ ptr = (int*)&room->light; }
-	else if(!str_cmp(field,"sector"))		{ ptr = (int*)&room->sector_type; flags = sector_flags; }
-	else if(!str_cmp(field,"heal"))			{ ptr = (int*)&room->heal_rate; min_sec = 9; }
-	else if(!str_cmp(field,"mana"))			{ ptr = (int*)&room->mana_rate; min_sec = 9; }
-	else if(!str_cmp(field,"move"))			{ ptr = (int*)&room->move_rate; min_sec = 1; }
-	else if(!str_cmp(field,"mapx"))			{ ptr = (int*)&room->x; min_sec = 5; allow_static = false; }
-	else if(!str_cmp(field,"mapy"))			{ ptr = (int*)&room->y; min_sec = 5; allow_static = false; }
-	else if(!str_cmp(field,"rsflags"))		{ ptr = (int*)room->rs_room_flag; bank = room_flagbank; allow_static = false; }
-	else if(!str_cmp(field,"rssector"))		{ ptr = (int*)&room->rs_sector_type; allow_static = false; }
-	else if(!str_cmp(field,"rsheal"))		{ ptr = (int*)&room->rs_heal_rate; min_sec = 9; allow_static = false; }
-	else if(!str_cmp(field,"rsmana"))		{ ptr = (int*)&room->rs_mana_rate; min_sec = 9; allow_static = false; }
-	else if(!str_cmp(field,"rsmove"))		{ ptr = (int*)&room->rs_move_rate; min_sec = 1; allow_static = false; }
-	else if(!str_cmp(field,"tempstore1"))	{ ptr = (int*)&room->tempstore[0]; }
-	else if(!str_cmp(field,"tempstore2"))	{ ptr = (int*)&room->tempstore[1]; }
-	else if(!str_cmp(field,"tempstore3"))	{ ptr = (int*)&room->tempstore[2]; }
-	else if(!str_cmp(field,"tempstore4"))	{ ptr = (int*)&room->tempstore[3]; }
+	if(!str_cmp(field,"flags"))				{ ptr = (long*)room->room_flag; bank = room_flagbank; }
+	else if(!str_cmp(field,"light"))		{ ptr = (long*)&room->light; }
+	else if(!str_cmp(field,"sector"))		{ ptr = (long*)&room->sector_type; flags = sector_flags; }
+	else if(!str_cmp(field,"heal"))			{ ptr = (long*)&room->heal_rate; min_sec = 9; }
+	else if(!str_cmp(field,"mana"))			{ ptr = (long*)&room->mana_rate; min_sec = 9; }
+	else if(!str_cmp(field,"move"))			{ ptr = (long*)&room->move_rate; min_sec = 1; }
+	else if(!str_cmp(field,"mapx"))			{ ptr = (long*)&room->x; min_sec = 5; allow_static = false; }
+	else if(!str_cmp(field,"mapy"))			{ ptr = (long*)&room->y; min_sec = 5; allow_static = false; }
+	else if(!str_cmp(field,"rsflags"))		{ ptr = (long*)room->rs_room_flag; bank = room_flagbank; allow_static = false; }
+	else if(!str_cmp(field,"rssector"))		{ ptr = (long*)&room->rs_sector_type; allow_static = false; }
+	else if(!str_cmp(field,"rsheal"))		{ ptr = (long*)&room->rs_heal_rate; min_sec = 9; allow_static = false; }
+	else if(!str_cmp(field,"rsmana"))		{ ptr = (long*)&room->rs_mana_rate; min_sec = 9; allow_static = false; }
+	else if(!str_cmp(field,"rsmove"))		{ ptr = (long*)&room->rs_move_rate; min_sec = 1; allow_static = false; }
+	else if(!str_cmp(field,"tempstore1"))	{ ptr = (long*)&room->tempstore[0]; }
+	else if(!str_cmp(field,"tempstore2"))	{ ptr = (long*)&room->tempstore[1]; }
+	else if(!str_cmp(field,"tempstore3"))	{ ptr = (long*)&room->tempstore[2]; }
+	else if(!str_cmp(field,"tempstore4"))	{ ptr = (long*)&room->tempstore[3]; }
 
 	if(!ptr && !sptr) return;
